powerpc64le: link tests against ld.so
As noted in the following commit

commit 91ac3a7d8474480685632cd25f844d3154c69fdf
Author: Tulio Magno Quites Machado Filho <tuliom@linux.vnet.ibm.com>
Date:   Mon Jul 17 17:48:59 2017 -0300

    powerpc: Fix float128 IFUNC relocations [BZ #21707]

libgcc depends on a symbol, __parse_hwcap_and_convert_at_platform,
provided by the loaders.  This dependency reflected on the need to link
some float128 tests against the loader.

In preparation for the transition of the format of long double - from
IBM Extended Precision to IEEE 754 128-bits floating-point - on
powerpc64le, this patch adds the linking of the loader to the tests for
long double, since after the switch they will also depend on
__parse_hwcap_and_convert_at_platform.

Tested on powerpc64le.