[RFC] Remove -ffloat-store from math/Makefile
Rationale:

The use of -ffloat-store causes some ICEs.  This has been discussed with
the compiler team and they explained to me that it is rather weird that
-ffloat-store is being used on POWER (-float-store is a x86 flag).

I don't know if this is a problem in glibc (which could be fixed in
glibc itself) or if it is an actual problem in the compiler (which would
required a bug report on bugzilla).