ldbl-128ibm-compat: Provide nexttoward functions
The functions in the nexttoward family are special, in the sense that
they always have a long double argument, regardless of their suffix
(i.e.: nexttowardf and nexttoward have a long double argument, besides
the float and double arguments).

On top of that, they are also special because nexttoward functions are
not part of the _FloatN API, hence __nexttowardf128 do not exist.

This patch adds 4 new function implementations for the new long double
format:

  __nexttoward_to_ieee128
  __nexttowardf_to_ieee128
  __nexttowardieee128 (as an alias to __nextafterieee128)

2018-06-19  Gabriel F. T. Gomes  <gabriel@inconstante.eti.br>
	    Tulio Magno Quites Machado Filho  <tuliom@linux.ibm.com>

	* sysdeps/ieee754/ldbl-128ibm-compat/Versions: Add
	__nexttowardf_to_ieee128, __nexttowardieee128 and
	__nexttoward_to_ieee128.
	* sysdeps/ieee754/ldbl-128ibm-compat/s_nextafterf128.c: New file.
	* sysdeps/ieee754/ldbl-128ibm-compat/s_nexttowardf128.c: New file.