Refactor math/bits/math-finite.h, reusing math/bits/mathcalls.h [BZ #23292]
On a new enough compiler, redirecting the same function twice can cause
-Werror=pragmas errors.  This issue may appear when redirecting a
floating point function to a different ABI at the same time that finite
redirections are expected.

This patch removes all of the common redirections done in math-finite.h
in order to reuse mathcalls.h.  Prototypes and redirections are now
defined only once.  The finite functions are identified with macros
__MATHCALL_FINITE or __MATHCALL_VEC_FINITE.
When finite functions are not requested, the macros behave in the same
way as their non-finite peers, i.e. __MATHCALL and __MATHCALL_VEC.
However, when finite functions are requested, i.e.
__HAVE_FINITE == 1, they both redirect their functions via
__MATHREDIR, using __MATH_FINITE_NAME to set the name of the target
function.

There is a new header called bits/math-features.h that indicates whether
the machine and the current compile unit are redirected to finite
functions (via __HAVE_FINITE).

In another change, math-finite.h inline functions are modified in order
to benefit from functions that have already been redirected.
Macro __REDIRTO_PUBLIC_F is used to reuse functions that are already
available in the API, e.g. lgammal is is an inline function that calls
__lgammal_r_finite, which has already been redirected via assembly to
__lgammaf128_r_finite.  The inline function could be directly redirected
to its respective f128 as it isn't guaranteed to be available in the API.

2018-06-20  Tulio Magno Quites Machado Filho  <tuliom@linux.ibm.com>
	    Rajalakshmi Srinivasaraghavan  <raji@linux.vnet.ibm.com>

	[BZ #23292]
	* bits/math-features.h: New file.  Provide __HAVE_FINITE.
	* sysdeps/ia64/fpu/bits/math-features.h: New file disabling
	__HAVE_FINITE.
	* math/bits/math-finite.h: Move most of the functions to
	bits/mathcalls.h.
	* math/bits/mathcalls.h: Use macros __MATHCALL_FINITE and
	__MATHCALL_VEC_FINITE to identify finite functions.
	* math/Makefile (headers): Add bits/math-features.h.
	* math/math.h: Include bits/math-features.h and replace all
	usages of "defined __FINITE_MATH_ONLY__ && __FINITE_MATH_ONLY__ > 0"
	with __HAVE_FINITE.
	(__MATH_FINITE_NAMEY, __MATH_FINITE_NAMEX,
	__MATH_FINITE_NAME_IMPL, __MATH_FINITE_NAME,
	__MATHCALL_VEC_FINITE, __MATHCALL_FINITE):
	New macros.
	[__HAVE_FINITE]: Include bits/math-finite.h for double, float
	and long double earlier in order to use its functions from the
	math-finite.h of _FloatN types.